Add `xsv partition` to partition on a column's values
This patch adds a new `xsv partition` subcommand, and makes a few minor
modifications to `xsv split` for consistency.  The new subcommand is
used as follows:

    xsv partition state out/ cities.csv

This will take the data in `cities.csv` and use the `state` column to
create new file names, creating files like:

    out/CA.csv
    out/TX.csv

If the `states` column is empty, we'll put the data into:

    out/empty.csv

There's an option for specifying the filename template:

    xsv partition --filename=state-{}.csv state out/ cities.csv

This will create:

    out/state-CA.csv
    out/state-TX.csv

This `--filename` option may also be used for `xsv split`.

There's also a `--prefix-length` argument, which can be used to limit
the number of files created:

    xsv partition --prefix-length=3 zip out/ customers.csv

This will create files using the first three digits of the zip code:

    out/000.csv
    out/010.csv
    out/011.csv
    out/213.csv

Note that if you try to split into more than roughly 500 files on
Windows or 1000 files on Linux systems, you may need to adjust `ulimit`
or the local equivalent.  (This was discussed in the original proposal.)

You can also write `--filename={}/file.csv`.  This supports two use
cases:

    xsv partition --filename {}/cities.csv state . all-cities.csv
    xsv partition --filename {}/monuments.csv state . all-monuments.csv
    xsv partition --filename {}/parks.csv state . all-parks.csv

Above, we want to partition our records by state into separate
directories, but we have multiple kinds of data.

    xsv partition --filename {}/$(uuidgen).csv . input1.csv
    xsv partition --filename {}/$(uuidgen).csv . input2.csv

Above, we're running multiple (possibly parallel) copies of xsv and we
want to parition the data into multiple directories without a filename
clash.

Fixes #51.