Add type parameters to from_bytes calls
There is a pull request on the rust compiler (PR #41336) that will add
support for a += &b (instead of only a += b) for numeric types like i64
and f64.

Unfortunately, this breaks the build of xsv, because xsv currently does
the following in a few places:

*float += from_bytes(sample).unwrap();

These lines rely on type inference to infer that from_bytes(...) must be
from_bytes::<f64>(...), but if the PR on the compiler is accepted, this
inference is no longer possible because from_bytes::<&'a f64>(...) would
also be valid.

This commit adds explicit type parameters to all from_bytes calls.