small performance improvements
For the most part, this switches write_record to write_byte_record where
possible, and also moves some uses of the `byte_records` iterator to the
manual `read_byte_record` method, which permits us to amortize
allocation.

This also removes the use of normal selection for `xsv search`, which
was causing a huge slow down.

Also, bump to csv 1.0.0.beta.3.