cmd/fmt: recognize tsv automatically
All commands except for `xsv fmt` recognized tsv files automatically,
and we seem to do this by surreptitiously using `(default: ,)` instead
of `[default: ,]` in our docopt description. The former does nothing,
while the latter actually specifies a default value. We want the former
so that our reader configuration knows when to try to be smart, which it
won't be if it's given an explicit delimiter.

Fixes #112