Add "--align" flag to table command.
This flag allows the user to specify how data should be aligned within
columns. The options are:

   * left
   * right
   * center

This alignment is applied to all columns.

Per the tabwriter crate, center alignment breaks ties in favor of being
closer to the left. Thus if "hi" appears in a column of width 5, it will
be printed as " hi  ".

resolves #165.