tests: fix BOM issue
Basically, the problem here is that the quickcheck tests will sometimes
generate CSV data that begins with a UTF-8 BOM and the CSV reader
automatically strips it.

There's a better way to solve this by making tests more robust, but they
were written without this consideration. As a hack, we just make sure
that we don't generate CSV data that begins with a BOM.

This isn't quite 100% correct since we could still get CSV data
beginning with a UTF-8 BOM through shrinking, but that only occurs when
a legitimate test failure happens. So this could potentially make
failure modes worse, but we abide for now.

Fixes #227